move method to gradle parser to new module
