+++ lib/salus/scanners/osv/gradle_osv.rb
@@ -3,6 +3,7 @@ require 'salus/scanners/osv/base'
 module Salus::Scanners::OSV
   class GradleOSV < Base
     class SemVersion < Gem::Version; end
+    include Gradle
 
     EMPTY_STRING = "".freeze
     DEFAULT_SOURCE = "https://osv.dev/list".freeze
